requirements.txt
slackclient
boto3